Moved functions to a new file, created requirement and readme.
