+++ requirements.txt
@@ -1,0 +1,10 @@
+numpy
+tulipy 
+schedule
+pyotp
+robin-stocks
+pyrh
+pytz
+python-dotenv
+yahoo_finance
+unittest
